Use 64-bit function table
Improves codegen for direct branches:

    mov qword [rax+0x408],0x10603560
 -  mov rcx,sub_10603560_OFFSET
 -  mov ecx,[rcx]
 -  mov ecx,ecx
 -  mov rdx,JIT_CACHE_BASE
 -  add rdx,rcx
 +  mov rcx,sub_10603560
 +  mov rdx,[rcx]
    mov rcx,rax

Improves codegen for dispatch stub:

    and rax,byte +0x1f
 -  mov eax,[rcx+rax*4]
 -  mov eax,eax
 -  mov rcx,JIT_CACHE_BASE
 -  lea rax,[rcx+rax]
 +  mov rax,[rcx+rax*8]
    mov rcx,rbx

+++ ARMeilleure/Translation/PTC/Ptc.cs
@@ -512,7 +512,7 @@ namespace ARMeilleure.Translation.PTC
             ConcurrentDictionary<ulong, TranslatedFunction> funcs,
             IMemoryManager memory,
             EntryTable<uint> countTable,
-            AddressTable<uint> funcTable,
+            AddressTable<ulong> funcTable,
             TranslatorStubs stubs)
         {
             if (AreCarriersEmpty())
@@ -649,7 +649,7 @@ namespace ARMeilleure.Translation.PTC
             RelocEntry[] relocEntries,
             IntPtr pageTablePointer,
             EntryTable<uint> countTable,
-            AddressTable<uint> funcTable,
+            AddressTable<ulong> funcTable,
             TranslatorStubs stubs,
             out Counter<uint> callCounter)
         {
@@ -779,7 +779,7 @@ namespace ARMeilleure.Translation.PTC
             ConcurrentDictionary<ulong, TranslatedFunction> funcs,
             IMemoryManager memory,
             EntryTable<uint> countTable,
-            AddressTable<uint> funcTable,
+            AddressTable<ulong> funcTable,
             TranslatorStubs stubs)
         {
             var profiledFuncsToTranslate = PtcProfiler.GetProfiledFuncsToTranslate(funcs);
